backend/requirements.txt
Flask~=2.0.2
dlib~=19.23.1
opencv-python~=4.5.2.52
numpy~=1.19.5
jsonpickle~=2.2.0
face_recognition~=1.3.0
scikit-learn~=0.24.2
scipy~=1.6.3
